refactor: 출석 상태 판단 로직을 Attendance의 책임으로 수정
