+++ src/test/java/domain/AttendanceTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
@@ -53,4 +54,119 @@ class AttendanceTest {
         assertThat(localTime).isEqualTo(LocalTime.of(10, 6));
     }
 
+    @Test
+    void 월요일_출석_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 16, 13, 0)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
+    }
+
+    @Test
+    void 월요일_출석_경계값_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 16, 13, 5)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
+    }
+
+    @Test
+    void 월요일_지각_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 16, 13, 6)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
+    }
+
+    @Test
+    void 월요일_지각_경계값_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 16, 13, 30)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
+    }
+
+    @Test
+    void 월요일_결석_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 16, 13, 31)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("결석");
+    }
+
+    @Test
+    void 화요일_출석_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 17, 10, 0)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
+    }
+
+    @Test
+    void 화요일_출석_경계값_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 17, 10, 5)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
+    }
+
+    @Test
+    void 화요일_지각_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 17, 10, 6)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
+    }
+
+    @Test
+    void 화요일_지각_경계값_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 17, 10, 30)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
+    }
+
+    @Test
+    void 화요일_결석_테스트() {
+        Attendance attendance = new Attendance(
+                LocalDateTime.of(2024, 12, 17, 10, 31)
+        );
+
+        AttendanceStatus attendanceStatus = attendance.judge();
+
+        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("결석");
+    }
+
+    @Test
+    void 주말의_경우_예외를_발생시킨다() {
+        assertThatThrownBy(() ->
+                new Attendance(LocalDateTime.of(2024, 12, 28, 10, 0)));
+    }
 }
